#Agnostic Bayes Ensemble Documentation

Overview

This package has been developed to facilitate increased predictive performance, by combining raw base models in an agnostic fashion, i.e. the methods don’t use any assumption regarding the used raw models. Furthermore, we specifically implemented ensemble algorithms that can deal with arbitrary loss function and with regression and classification problems, this holds true for all, except for the dirichletPosteriorEstimation algorithm, which is limited to classification problems.

Hint: In most cases it is advisable to deactivate Hyperthreading for best performance. However, in some rare cases – depending on the (hardware) platform the package runs on- you will get the best performance with Hyperthreading enabled, to be sure, it is best practice to measure the performance with and without Hyperthreading.

List of Algorithms

AgnosticBayesEnsemble.bootstrapPosteriorEstimation — Function

bootstrapPosteriorEstimation( errMat::Matrix{Float64}, samplingBatchSize::Int64, nrRuns::Int64 )

compute posterior p( h* = h | S ).
#Arguments
- `errMat::Matrix{Float64}}`: each column is the prediction error of one hypothesis.
- `samplingBatchSize::Int64`: sample size per main iteration.
- `nrRuns::Int64`:            number of passes over predictions.
#Return
- `Vector{Float64}`:          Distribution p( h* = h | S ).

AgnosticBayesEnsemble.bootstrapPosteriorEstimation! — Function

bootstrapPosteriorEstimation!( errMat::Matrix{Float64}, samplingBatchSize::Int64, nrRuns::Int64, p::Array{Float64} )

compute posterior p( h* = h | S ).
#Arguments
- `errMat::Matrix{Float64}}`: each column is the prediction error of one hypothesis.
- `samplingBatchSize::Int64`: sample size per main iteration.
- `nrRuns::Int64`:            number of passes over predictions.
- `p::Vector{Float64}`:       resulting posterior p( h* = h | S ).
#Return
- `nothing`:                  nothing.

AgnosticBayesEnsemble.bootstrapPosteriorCorEstimation — Function

bootstrapPosteriorCorEstimation( predictions::Matrix{Float64}, y::Array{Float64,1}, samplingBatchSize::Int64, nrRuns::Int64 )

compute posterior p( h* = h | S ).
#Arguments
- `predictions::Matrix{Float64}`: each column is the prediction error of one hypothesis.
- `t::Vector{Float64}`:           label vector.
- `samplingBatchSize::Int64`:     sample size per main iteration.
- `nrRuns::Int64`:                number of main  iterations.
#Return
- `Float64`:            Best found meta parameter α.

bootstrapPosteriorCorEstimation( predictions::Matrix{Float64}, y::Array{Float64,1}, samplingBatchSize::Int64, nrRuns::Int64 )

compute posterior p( h* = h | S ).
#Arguments
- `predictions::Matrix{Float64}`: each column is the prediction error of one hypothesis.
- `T::Vector{Float64}`:           label matrix.
- `samplingBatchSize::Int64`:     sample size per main iteration.
- `nrRuns::Int64`:                number of main  iterations.
#Return
- `Float64`:            Best found meta parameter α.

AgnosticBayesEnsemble.dirichletPosteriorEstimation — Function

dirichletPosteriorEstimation( errMat::Matrix{Float64}, G::Matrix{Float64}, nrRuns::Int64, α_ )

compute posterior p( h* = h | S ).
# Arguments
- `errMat::Matrix{Float64}`: each column is the prediction of one hypothesis.
- `nrRuns::Int64`:           number of sampling runs.
- `α_::Float64`:             scalar prior parameter.
- `sampleSize::Int64`:       number of samples per run.
# Return
- `Vector{Float64}`:         posterior distribution

dirichletPosteriorEstimation( errMat::Matrix{Float64}, nrRuns::Int64, α_::Float64 )

compute posterior p( h* = h | S ).
# Arguments
- `errMat::Matrix{Float64}`: each column is the prediction of one hypothesis.
- `nrRuns::Int64`:           number of sampling runs.
- `α_::Float64`:             scalar prior parameter.
- `sampleSize::Int64`:       number of samples per run.
# Return
- `Vector{Float64}`:         posterior distribution

AgnosticBayesEnsemble.dirichletPosteriorEstimationV2 — Function

dirichletPosteriorEstimationV2( errMat::Matrix{Float64}, nrRuns::Int64, α_::Float64, sampleSize::Int64 )
compute posterior p( h* = h | S ).
# Arguments
- `errMat::Matrix{Float64}`: each column is the prediction of one hypothesis.
- `nrRuns::Int64`:           number of sampling runs.
- `α_::Float64`:             scalar prior parameter.
- `sampleSize::Int64`:       number of samples per run.
# Return
- `Vector{Float64}`:         posterior distribution

AgnosticBayesEnsemble.GMatrix — Function

GMatrix( d::Int64 )

compute transformation matrix G.
#Arguments
- `d::Int64`:        number of hypothesis used for prediction.
#Return
- `Matrix{Float64}`: transformation matrix G.

AgnosticBayesEnsemble.dirichletPosteriorEstimation! — Function

dirichletPosteriorEstimation!( errMat::Matrix{Float64}, nrRuns::Int64, α_::Float64, p::Vector{Float64} )

compute posterior p( h* = h | S ).
#Arguments
- `errMat::Matrix{Float64}`: each column is the prediction error of one hypothesis.
- `nrRuns::Int64`:           number of passes over predictions.
- `α_::Float64`:             meta parameter value.
- `p::Vector{Float64}`:      return value posterior p( h* = h | S ).
#Return
- `Float64`:            Best found meta parameter α.

AgnosticBayesEnsemble.metaParamSearchValidationDirichlet — Function

metaParamSearchValidationDirichlet( Y::Matrix{Float64}, t::Vector{Float64}, nrRuns::Int64, minVal::Float64, maxVal::Float64, nSteps::Int64, holdout::Float64, lossFunc )    

compute best α parameter regarding predictive performance.
#Arguments
- `Y::Matrix{Float64}`: each column is the prediction error of one hypothesis.
- `t::Vector{Float64}`: label vector.
- `nrRuns::Int64`:      number of passes over predictions.
- `minVal::Float64`:    minimum value of α.
- `maxVal::Float64`:    maximum value of α.
- `nSteps::Int64`:      number of steps between min and max val.
- `holdout::Float64`:   percentage used in holdout.
- `lossFunc`:           error function handle.
#Return
- `Float64`:            Best found meta parameter α.

Tutorials

Another Section

Index

AgnosticBayesEnsemble.GMatrix

AgnosticBayesEnsemble.bootstrapPosteriorCorEstimation

AgnosticBayesEnsemble.bootstrapPosteriorEstimation

AgnosticBayesEnsemble.bootstrapPosteriorEstimation!

AgnosticBayesEnsemble.dirichletPosteriorEstimation

AgnosticBayesEnsemble.dirichletPosteriorEstimation!

AgnosticBayesEnsemble.dirichletPosteriorEstimationV2

AgnosticBayesEnsemble.metaParamSearchValidationDirichlet